backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.22.0
pydantic==2.12.3
requests==2.32.5
python-dotenv==1.2.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.4
openai==1.1.0
sqlalchemy==2.0.30



